move logic from handler files to alpacaService2
