+++ src/pages/api/broker/delete-order.ts
@@ -1,40 +1,9 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import fetch, { Response } from "node-fetch";
-import { handleResult } from "../../../util";
+import { Response } from "node-fetch";
+import { deleteOrder } from "../../../services/alpacaService";
 
 type Data = {
   data: Response;
-};
-
-const {
-  ALPACA_API_KEY_ID = "[NOT_DEFINED_IN_ENV]",
-  ALPACA_API_KEY_VALUE = "[NOT_DEFINED_IN_ENV]",
-} = process.env;
-
-const buff = Buffer.from(
-  `${ALPACA_API_KEY_ID}:${ALPACA_API_KEY_VALUE}`,
-  "utf-8",
-);
-const base64EncodedKeys = buff.toString("base64");
-
-const accountId = "b75acdbc-3fb6-3fb3-b253-b0bf7d86b8bb"; // public info
-const baseUrl = "https://broker-api.sandbox.alpaca.markets/v1";
-
-const deleteOrder = async (orderId: string) => {
-  try {
-    const res = await fetch(
-      `${baseUrl}/trading/accounts/${accountId}/orders/${orderId}`,
-      {
-        method: "DELETE",
-        headers: {
-          Authorization: `Basic ${base64EncodedKeys}`,
-        },
-      },
-    );
-    return await handleResult(res);
-  } catch (e) {
-    throw Error("Unable to delete order");
-  }
 };
 
 export default async function handler(
