+++ src/pages/api/broker/get-orders.ts
@@ -1,39 +1,9 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import fetch, { Response } from "node-fetch";
-import { handleResult } from "../../../util";
+import { Response } from "node-fetch";
+import { getOrders } from "../../../services/alpacaService";
 
 type Data = {
   orders: Response;
-};
-
-const {
-  ALPACA_API_KEY_ID = "[NOT_DEFINED_IN_ENV]",
-  ALPACA_API_KEY_VALUE = "[NOT_DEFINED_IN_ENV]",
-} = process.env;
-
-const buff = Buffer.from(
-  `${ALPACA_API_KEY_ID}:${ALPACA_API_KEY_VALUE}`,
-  "utf-8",
-);
-const base64EncodedKeys = buff.toString("base64");
-
-const accountId = "b75acdbc-3fb6-3fb3-b253-b0bf7d86b8bb"; // public info
-const baseUrl = "https://broker-api.sandbox.alpaca.markets/v1";
-
-const getOrders = async () => {
-  try {
-    const res = await fetch(
-      `${baseUrl}/trading/accounts/${accountId}/orders?status=all`,
-      {
-        headers: {
-          Authorization: `Basic ${base64EncodedKeys}`,
-        },
-      },
-    );
-    return await handleResult(res);
-  } catch (e) {
-    throw Error(`Unable to get orders - ${e as string}`);
-  }
 };
 
 export default async function handler(
